mds/requirements.txt
Fiona
geopandas
jsonschema
numpy
requests
pandas
psycopg2-binary
requests
scipy
Shapely
sqlalchemy